+++ Assets/Scripts/MB/ActorController.cs
@@ -29,7 +29,7 @@ public class ActorController : NeolithicObject {
 	public BaseOrder currentOrder = null;
 	public bool idle = false;
     public bool followContours = true;
-    public decimal health = 1.0m;
+    public double health = 1.0;
 
     public GameObject palzy;//temp debugging object
     public int queueLength = 0;
@@ -55,7 +55,7 @@ public class ActorController : NeolithicObject {
     }
 
 	public virtual void FixedUpdate() {
-        decimal feedMe = (decimal)(Time.fixedDeltaTime * 0.025f);
+        double feedMe = (double)(Time.fixedDeltaTime * 0.025f);
         LogisticsNetwork network = logisticsActor.logisticsManager.FindNearestNetwork(transform.position);
 
         if (network != null && network.foodbuffer > feedMe) {
@@ -65,7 +65,7 @@ public class ActorController : NeolithicObject {
             health -= feedMe;
         }
 
-        if (health <= 0.0m) {
+        if (health <= 0.0) {
             Destroy(gameObject);
             return;
         }
